python/requirements.txt
aniso8601==7.0.0
asyncpg==0.24.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
databases==0.5.3
Faker==9.8.0
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-GraphQL==2.0.1
Flask-GraphQL-Auth==1.3.3
graphene==2.1.9
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
marshmallow==3.14.0
promise==2.3
psycopg2==2.9.1
pycparser==2.21
PyJWT==2.0.1
python-dateutil==2.8.2
python-decouple==3.5
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
SQLAlchemy==1.4.26
text-unidecode==1.3
Werkzeug==2.0.2
